precompiler to TS
